requirements.txt
PyYAML~=5.3.1
w1thermsensor~=1.3.0
APScheduler~=3.6.3
gpiozero~=1.5.1